// packages/prisma-oppulence/generated/zod/inputTypeSchemas/AutomationTriggerMaxOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';

export const AutomationTriggerMaxOrderByAggregateInputSchema: z.ZodType<Prisma.AutomationTriggerMaxOrderByAggregateInput> = z.object({
  id: z.lazy(() => SortOrderSchema).optional(),
  automationId: z.lazy(() => SortOrderSchema).optional(),
  field: z.lazy(() => SortOrderSchema).optional(),
  operator: z.lazy(() => SortOrderSchema).optional(),
  value: z.lazy(() => SortOrderSchema).optional()
}).strict();

export default AutomationTriggerMaxOrderByAggregateInputSchema;